requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
pymysql==1.0.3
SQLAlchemy==1.4.49
Werkzeug==2.2.3
python-dotenv==1.0.0
cryptography==41.0.5 